requirements.api.txt
# Requirements for TASK-005: API Extensions
# FastAPI Application with 4 Endpoints
# + VICTOR BOT v2.0 Universal Sensor

# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0


# HTTP Client
httpx==0.26.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
python-multipart==0.0.6

# WebSocket
websockets==12.0

# Database
supabase==2.3.4
psycopg2-binary==2.9.9
asyncpg==0.29.0

# System Monitoring
psutil==5.9.7

# Prometheus Metrics (optional)
prometheus-client==0.19.0

# Async Support
aiohttp==3.9.1
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Logging
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0


# JWT Authentication
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# ============================================================================
# VICTOR BOT v2.0 - Universal Sensor Dependencies
# ============================================================================

# Telegram Bot
python-telegram-bot==20.7
aiogram==3.3.0

# OCR (Tesseract)
pytesseract==0.3.10
Pillow==10.2.0

# OpenAI (Whisper + GPT-4 Vision)
openai==1.7.2

# AWS S3 (File Storage)
boto3==1.34.20

# Face Recognition (optional)
# face-recognition==1.3.0
# opencv-python==4.9.0.80

# Table Extraction (optional)
# tabula-py==2.9.0
# camelot-py==0.11.0

# Background Tasks
# celery==5.3.4
# redis==5.0.1

# Redis for caching and queues
redis>=5.0.0

# Rate Limiting
slowapi==0.1.9
slowapi>=0.1.9

# ============================================================================
# Phase 6: Advanced ML Dependencies
# ============================================================================

# OpenAI Embeddings API
openai>=1.12.0

# Machine Learning
scikit-learn>=1.4.0
numpy>=1.26.0

# NLP & Sentiment Analysis
textblob>=0.18.0

# Job Scheduling
apscheduler>=3.10.4
apscheduler>=3.10.4

# GraphQL Server (Phase 6)
strawberry-graphql[fastapi]>=0.219.0
